Replace Vedeu::Models::Escape with Vedeu::Cells::Escape.
