+++ lib/vedeu/cells/escape.rb
@@ -17,6 +17,53 @@ module Vedeu
     #
     class Escape < Vedeu::Cells::Empty
 
+      include Vedeu::Geometries::Positionable
+
+      # @return [Boolean]
+      def cell?
+        false
+      end
+
+      # @return [NilClass]
+      def null
+        nil
+      end
+      alias_method :background, :null
+      alias_method :colour,     :null
+      alias_method :foreground, :null
+      alias_method :style,      :null
+
+      # Return an empty hash as most escape sequences won't make
+      # sense as JSON.
+      #
+      # @return [Hash<void>]
+      def to_h
+        {}
+      end
+      alias_method :to_hash, :to_h
+
+      # Return an empty string as most escape sequences won't make
+      # sense as HTML.
+      #
+      # @param _options [Hash] Ignored.
+      # @return [String]
+      def to_html(_options = {})
+        ''.freeze
+      end
+
+      # @return [String]
+      def to_s
+        "#{position}#{value}".freeze
+      end
+      alias_method :to_str, :to_s
+
+      private
+
+      # @return [Hash]
+      def defaults
+        super.merge!(position: [1, 1], value: '')
+      end
+
     end # Escape
 
   end # Cells
