+++ lib/vedeu/models/all.rb
@@ -10,7 +10,6 @@ end # Vedeu
 
 require 'vedeu/models/toggleable'
 require 'vedeu/models/cell'
-require 'vedeu/models/escape'
 require 'vedeu/models/focus'
 require 'vedeu/models/row'
 require 'vedeu/models/page'
